4_audio_anno_chat/requirements.txt
Flask
openai-whisper
# PyTorch ist die Basis-Bibliothek fuer Whisper und wird fuer die Berechnungen benoetigt
torch